# Port mdmd from mainline + Qwen2/2.5-VL support