--- a/ggml/src/ggml-cuda/conv2d-dw.cuh
+++ b/ggml/src/ggml-cuda/conv2d-dw.cuh
@@ -1,0 +1,5 @@
+#pragma once
+#include "common.cuh"
+
+#define CUDA_CONV2D_DW_BLOCK_SIZE 256
+void ggml_cuda_op_conv2d_dw(ggml_backend_cuda_context & ctx, ggml_tensor * dst);