… examples/CMakeLists.txt
… examples/main/main.cpp
… examples/mtmd/CMakeLists.txt
… examples/mtmd/README.md
… examples/mtmd/clip-impl.h
… examples/mtmd/clip.cpp
… examples/mtmd/clip.h
… examples/mtmd/deprecation-warning.cpp
… examples/mtmd/legacy-models/convert_image_encoder_to_gguf.py
… examples/mtmd/legacy-models/glmedge-convert-image-encoder-to-gguf.py
… examples/mtmd/legacy-models/glmedge-surgery.py
… examples/mtmd/legacy-models/llava_surgery.py
… examples/mtmd/legacy-models/llava_surgery_v2.py
… examples/mtmd/legacy-models/minicpmv-convert-image-encoder-to-gguf.py
… examples/mtmd/legacy-models/minicpmv-surgery.py
… examples/mtmd/mtmd-audio.cpp
… examples/mtmd/mtmd-audio.h
… examples/mtmd/mtmd-cli.cpp
… examples/mtmd/mtmd-helper.cpp
… examples/mtmd/mtmd-helper.h
… examples/mtmd/mtmd.cpp
… examples/mtmd/mtmd.h
… examples/mtmd/requirements.txt
… examples/mtmd/test-1.jpeg
… examples/mtmd/test-2.mp3
… examples/mtmd/tests.sh
… examples/server/server.cpp
… ggml/src/ggml-cuda/conv2d.cu
… ggml/src/ggml-cuda/convert.cuh
… ggml/src/ggml-cuda/rope.cu