@ikawrakow 2025-09-27
Nothing has been done for the server. You have a command line tool. You can use it like this
```
./bin/llama-mtmd-cli -m model --mmproj mmproj $other_args
...
> /image some_image.jpg
> Describe the image
...
```
(see the `mtmd` documentation)

For the server I'm hoping someone else will do a PR.