@ikawrakow 2025-09-28
> mmproj-f16.gguf is acepted but not mmproj-Q8_0.gguf

Does mainline support quantized mmproj files? To me it looked like the convolution and `im2col` kernels only work with `f16/f32` tensors.