@mcm007 2025-09-28
Yes, [mmproj-Qwen2.5-VL-7B-Instruct-Q8_0.gguf](https://huggingface.co/ggml-org/Qwen2.5-VL-7B-Instruct-GGUF/blob/main/mmproj-Qwen2.5-VL-7B-Instruct-Q8_0.gguf) produces good results in mainline; it seems it needs `--no-mmproj-offload` otherwise it produces some random text each time.